services/ai-service/ChatBot/requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-openai
langchain-community
langchain-text-splitters
langchainhub

chromadb
tiktoken

pymongo
python-dotenv
jinja2
requests
aiofiles
